An Algorithmic trading tutorial on how to use Exponential Moving Averages (EMA) & MACD to create trading signals to buy or sell a stock (or a commodity).
MACD is defined as Moving Average Convergence Divergence. It is the difference between a short term EMA and a Long Term EMA, analysed against a shorter term EMA to generate trading signals based on overall market behaviour and price momentum.

Markets are analogous to oscillating pendulums, whose amplitude depends on the forces applied to it. Forces can be seen as the pressure to buy or sell the product over a certain period. As long as the forces (buy or sell) are applied to the oscillating pendulum, the signal curve and the MACD curve will continue to cross each other and oscillate whenever a shift in trend occurs. This is the basis to understand the trading signals returned in this code.